correct loyalty_points field in Order model

+++ checkout/models.py
@@ -25,7 +25,7 @@ class Order(models.Model):
     grand_total = models.DecimalField(max_digits=10, decimal_places=2, null=False, default=0)
     original_cart = models.TextField(null=False, blank=False, default='')
     stripe_pid = models.CharField(max_length=254, null=False, blank=False, default='')
-    loyalty_points = models.DecimalField(max_digits=10, decimal_places=2, null=False, default=0)
+    loyalty_points = models.DecimalField(max_digits=10, decimal_places=0, null=False, default=0)
 
     def _generate_order_number(self):
         """
@@ -54,7 +54,7 @@ class Order(models.Model):
         if self.order_total < 10:
             self.loyalty_points = 0
         else:
-            self.loyalty_points = self.order_total / 10
+            self.loyalty_points = int(self.order_total / 10)
         self.save()
 
     def save(self, *args, **kwargs):
